Added security constants
